// admin-dashboard/src/app/services/models/page-response-registered-users.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RegisteredUsers } from '../models/registered-users';
export interface PageResponseRegisteredUsers {
  content?: Array<RegisteredUsers>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}