// admin-dashboard/src/app/services/models/registered-users.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegisteredUsers {
  createdAt?: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  publicId?: string;
  roles?: Array<'SUPER_ADMIN' | 'ADMIN' | 'DRIVER' | 'CONSIGNOR'>;
  status?: 'OTP_SENT' | 'VERIFIED' | 'PENDING' | 'APPROVED' | 'ACTIVE' | 'DISABLED' | 'SUSPENDED' | 'REJECTED';
}